Added in a bit that allows for some more data to automatically be acquired. I can't figure out how to get the BuildingDisplay.cs to do this (it still functions right now, but ideally I'd like to move it to the automatic system). Other than that, the stats screen also works. :)

+++ Assets/My Assets/Scripts/Building.cs
@@ -52,7 +52,7 @@ namespace buildingClass {
 			ulong cost = 0;
 
 			for (int i = 0; i < amount;  i++) {
-				cost += (ulong)(costForFirst * Mathf.Pow((float)getIncreasePerUnit(), getNum() + i));
+				cost += (ulong)(costForFirst * Mathf.Pow(getIncreasePerUnit(), getNum() + i));
             }
 
 			return cost;
